requirements.txt
aiohttp==3.8.4
aiojobs==0.2.2
async-timeout==4.0.2
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==8.0.0
decorator==4.4.1
google-auth
idna==2.9
iso8601==0.1.12
jsonpath-ng==1.4.3
kopf==1.36
kubernetes==17.17.0
multidict==4.7.5
oauthlib==3.1.0
ply==3.11
pyasn1==0.4.8
pyasn1-modules==0.2.8
pykube-ng==20.1.0
python-dateutil==2.8.1
PyYAML==5.4
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.26.5
websocket-client==0.57.0
yarl==1.8.2